+++ tests/test_ui_editor_rescue_surface.py
@@ -76,8 +76,8 @@ def test_legacy_joined_runtime_files_are_removed() -> None:
 
 def test_shell_layout_guards_prevent_duplicate_suggestions_and_forced_chat_settings() -> None:
     primary_runtime = _read_all_runtime_js()
-    layout_css = _read_text("thomas/server/web/css/layout_parts/layout-app-shell.css")
-    suggestion_css = _read_text("thomas/server/web/css/components_parts/easy-setup-ui.css")
+    layout_css = _read_text("thomas/server/web/css/layout_styles/layout-app-shell.css")
+    suggestion_css = _read_text("thomas/server/web/css/component_styles/easy-setup-ui.css")
     marketplace_css = _read_text("thomas/server/web/static/plugin_marketplace.style01_part01.css")
 
     open_settings_block = primary_runtime.split("function openSettingsModal()", 1)[1].split(